+++ src/main/java/top/sdugyf/community/community/dto/QuestionDTO.java
@@ -8,12 +8,12 @@ import top.sdugyf.community.community.model.User;
 public class QuestionDTO {
 
 
-    private Integer id;
+    private Long id;
     private String title;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
